// ui/scheduler_ui/src/api/models/scheduleSolveResponse.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Scheduler API
 * OpenAPI spec version: 1.0.0
 */
import type { SchedulerApiDbModelsEnumsScheduleSolveStatus } from './schedulerApiDbModelsEnumsScheduleSolveStatus';
import type { ScheduleSolveResponseParameters } from './scheduleSolveResponseParameters';

export interface ScheduleSolveResponse {
  id: string;
  schedule_template_id: string;
  schedule_id: string | null;
  status: SchedulerApiDbModelsEnumsScheduleSolveStatus;
  started_at: string | null;
  finished_at: string | null;
  parameters: ScheduleSolveResponseParameters;
  celery_task_id: string | null;
  current_generation: number | null;
  best_fitness: number | null;
  progress: number;
  error_details: string | null;
  created_at: string;
  updated_at: string;
}