// ui/scheduler_ui/src/api/models/schedulerApiDbModelsEnumsScheduleSolveStatus.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Scheduler API
 * OpenAPI spec version: 1.0.0
 */

export type SchedulerApiDbModelsEnumsScheduleSolveStatus = typeof SchedulerApiDbModelsEnumsScheduleSolveStatus[keyof typeof SchedulerApiDbModelsEnumsScheduleSolveStatus];


export const SchedulerApiDbModelsEnumsScheduleSolveStatus = {
  pending: 'pending',
  queued: 'queued',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
  cancelled: 'cancelled',
} as const;